@@ -14,25 +14,25 @@
         <Nullable>annotations</Nullable>
     </PropertyGroup>
     <PropertyGroup Condition="!$([MSBuild]::IsOSUnixLike())">
-        <TargetFrameworks>net452;netcoreapp3.1</TargetFrameworks>
+        <TargetFrameworks>net48;net6.0</TargetFrameworks>
     </PropertyGroup>
     <PropertyGroup Condition="$([MSBuild]::IsOSUnixLike())">
-        <TargetFramework>netcoreapp3.1</TargetFramework>
+        <TargetFramework>net6.0</TargetFramework>
     </PropertyGroup>
     <ItemGroup>
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.3.2" />
         <ProjectReference Include="..\Octopus.Tentacle\Octopus.Tentacle.csproj" />
-        <PackageReference Include="NSubstitute" Version="4.2.2" />
-        <PackageReference Include="FluentAssertions" Version="5.10.3" />
-        <PackageReference Include="Assent" Version="1.6.0" />
-        <PackageReference Include="NUnit" Version="3.12.0" />
-        <PackageReference Include="NUnit3TestAdapter" Version="3.13.0" />
+        <PackageReference Include="NSubstitute" Version="4.4.0" />
+        <PackageReference Include="FluentAssertions" Version="6.7.0" />
+        <PackageReference Include="Assent" Version="1.8.2" />
+        <PackageReference Include="NUnit" Version="3.13.3" />
+        <PackageReference Include="NUnit3TestAdapter" Version="4.3.0" />
         <PackageReference Include="System.ComponentModel.TypeConverter" Version="4.3.0" />
         <PackageReference Include="System.Runtime.InteropServices.RuntimeInformation" Version="4.3.0" />
         <PackageReference Include="TeamCity.VSTest.TestAdapter" Version="1.0.15" />
-        <Reference Include="System.DirectoryServices.AccountManagement" Condition="'$(TargetFramework)' == 'net452'" />
-        <Reference Include="System.IdentityModel" Condition="'$(TargetFramework)' == 'net452'" />
-        <Reference Include="System.ServiceProcess" Condition="'$(TargetFramework)' == 'net452'" />
+        <Reference Include="System.DirectoryServices.AccountManagement" Condition="'$(TargetFramework)' == 'net48'" />
+        <Reference Include="System.IdentityModel" Condition="'$(TargetFramework)' == 'net48'" />
+        <Reference Include="System.ServiceProcess" Condition="'$(TargetFramework)' == 'net48'" />
     </ItemGroup>
     <ItemGroup>
       <Content Include="Startup\Packages\Acme.Service\Acme.Service.exe">
