@@ -50,8 +50,7 @@ namespace Octopus.Tentacle.Diagnostics
             if (sensitiveValues == null || sensitiveValues.Length == 0)
                 return this;
 
-            var initialSensitiveValuesCount = SensitiveValues?.Length ?? 0;
-            // ReSharper disable once RedundantEnumerableCastCall
+            var initialSensitiveValuesCount = SensitiveValues.Length;
             var sensitiveValuesUnion = SensitiveValues.Union(sensitiveValues.Where(x => x != null).Cast<string>()).ToArray();
 
             // If no new sensitive-values were added, avoid the cost of rebuilding the trie
